Use the standard Quarkus extension project structure.
This should allow us to push the quarkus extension to quarkiverse eventually.
