--- proxy-wasm-jaxrs-quarkus/src/test/java/io/roastedroot/proxywasm/jaxrs/example/tests/FFITest.java
@@ -1,22 +1,0 @@
-package io.roastedroot.proxywasm.jaxrs.example.tests;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.equalTo;
-
-import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-
-@QuarkusTest
-public class FFITest {
-
-    @Test
-    public void reverse() throws InterruptedException {
-
-        given().body("My Test")
-                .when()
-                .post("/ffiTests/reverse")
-                .then()
-                .statusCode(200)
-                .body(equalTo("tseT yM"));
-    }
-}
