--- proxy-wasm-jaxrs-quarkus/src/test/java/io/roastedroot/proxywasm/jaxrs/example/tests/HttpCallTest.java
@@ -1,25 +1,0 @@
-package io.roastedroot.proxywasm.jaxrs.example.tests;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.equalTo;
-
-import io.quarkus.test.junit.QuarkusTest;
-import io.roastedroot.proxywasm.StartException;
-import org.junit.jupiter.api.Test;
-
-@QuarkusTest
-public class HttpCallTest {
-
-    @Test
-    public void test() throws InterruptedException, StartException {
-
-        // the wasm plugin will forward the request to the /ok endpoint
-        given().header("test", "ok")
-                .when()
-                .get("/httpCallTests")
-                .then()
-                .statusCode(200)
-                .body(equalTo("ok"))
-                .header("echo-test", "ok");
-    }
-}
