--- proxy-wasm-jaxrs-quarkus/src/test/resources/application.properties
@@ -1,2 +1,0 @@
-quarkus.log.level=INFO
-quarkus.log.category."org.hibernate".level=DEBUG
