+++ proxy-wasm-quarkus-example/src/main/java/io/roastedroot/proxywasm/jaxrs/example/App.java
