+++ proxy-wasm-quarkus-example/src/main/resources/application.properties
