+++ proxy-wasm-quarkus-example/src/test/java/io/roastedroot/proxywasm/jaxrs/example/FFITest.java
