+++ quarkus-proxy-wasm/.github/project.yml
@@ -1,0 +1,4 @@
+release:
+  current-version: "0"
+  next-version: "1.0-SNAPSHOT"
+
