+++ quarkus-proxy-wasm/.github/workflows/pre-release.yml
@@ -1,0 +1,16 @@
+name: Quarkiverse Pre Release
+
+on:
+  pull_request:
+    paths:
+      - '.github/project.yml'
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  pre-release:
+    name: Pre-Release
+    uses: quarkiverse/.github/.github/workflows/pre-release.yml@main
+    secrets: inherit
