+++ quarkus-proxy-wasm/.github/workflows/quarkus-snapshot.yaml
@@ -1,0 +1,55 @@
+name: "Quarkus ecosystem CI"
+on:
+  workflow_dispatch:
+  watch:
+    types: [started]
+
+  # For this CI to work, ECOSYSTEM_CI_TOKEN needs to contain a GitHub with rights to close the Quarkus issue that the user/bot has opened,
+  # while 'ECOSYSTEM_CI_REPO_PATH' needs to be set to the corresponding path in the 'quarkusio/quarkus-ecosystem-ci' repository
+
+env:
+  ECOSYSTEM_CI_REPO: quarkusio/quarkus-ecosystem-ci
+  ECOSYSTEM_CI_REPO_FILE: context.yaml
+  JAVA_VERSION: 17
+
+  #########################
+  # Repo specific setting #
+  #########################
+
+  ECOSYSTEM_CI_REPO_PATH: quarkiverse-proxy-wasm
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+defaults:
+  run:
+    shell: bash
+
+jobs:
+  build:
+    name: "Build against latest Quarkus snapshot"
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Set up Java
+        uses: actions/setup-java@v4
+        with:
+          distribution: temurin
+          java-version: ${{ env.JAVA_VERSION }}
+
+      - name: Checkout repo
+        uses: actions/checkout@v4
+        with:
+          path: current-repo
+
+      - name: Checkout Ecosystem
+        uses: actions/checkout@v4
+        with:
+          repository: ${{ env.ECOSYSTEM_CI_REPO }}
+          path: ecosystem-ci
+
+      - name: Setup and Run Tests
+        run: ./ecosystem-ci/setup-and-test
+        env:
+          ECOSYSTEM_CI_TOKEN: ${{ secrets.ECOSYSTEM_CI_TOKEN }}
