+++ quarkus-proxy-wasm/.github/workflows/release-prepare.yml
@@ -1,0 +1,19 @@
+name: Quarkiverse Prepare Release
+
+on:
+  workflow_dispatch:
+  pull_request:
+    types: [ closed ]
+    paths:
+      - '.github/project.yml'
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  prepare-release:
+    name: Prepare Release
+    if: ${{ github.event_name == 'workflow_dispatch' || github.event.pull_request.merged == true}}
+    uses: quarkiverse/.github/.github/workflows/prepare-release.yml@main
+    secrets: inherit
