+++ quarkus-proxy-wasm/deployment/src/main/java/io/quarkiverse/proxywasm/deployment/ProxyWasmProcessor.java
@@ -1,0 +1,31 @@
+package io.quarkiverse.proxywasm.deployment;
+
+import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+import io.quarkus.jaxrs.spi.deployment.AdditionalJaxRsResourceMethodAnnotationsBuildItem;
+import io.roastedroot.proxywasm.jaxrs.WasmPlugin;
+import io.roastedroot.proxywasm.jaxrs.cdi.WasmPluginFeature;
+import java.util.List;
+import org.jboss.jandex.DotName;
+
+class ProxyWasmProcessor {
+
+    private static final String FEATURE = "proxy-wasm";
+
+    @BuildStep
+    FeatureBuildItem feature() {
+        return new FeatureBuildItem(FEATURE);
+    }
+
+    @BuildStep
+    AdditionalBeanBuildItem resources() {
+        return new AdditionalBeanBuildItem(WasmPluginFeature.class);
+    }
+
+    @BuildStep
+    public AdditionalJaxRsResourceMethodAnnotationsBuildItem annotations() {
+        return new AdditionalJaxRsResourceMethodAnnotationsBuildItem(
+                List.of(DotName.createSimple(WasmPlugin.class.getName())));
+    }
+}
