+++ quarkus-proxy-wasm/docs/modules/ROOT/nav.adoc
@@ -1,0 +1,1 @@
+* xref:index.adoc[Getting started]
