+++ quarkus-proxy-wasm/docs/modules/ROOT/pages/includes/attributes.adoc
@@ -1,0 +1,3 @@
+:project-version: 0
+
+:examples-dir: ./../examples/
