+++ quarkus-proxy-wasm/docs/modules/ROOT/pages/index.adoc
@@ -1,0 +1,27 @@
+= Quarkus Proxy Wasm
+
+include::./includes/attributes.adoc[]
+
+TIP: Describe what the extension does here.
+
+== Installation
+
+If you want to use this extension, you need to add the `io.quarkiverse.proxy-wasm:quarkus-proxy-wasm` extension first to your build file.
+
+For instance, with Maven, add the following dependency to your POM file:
+
+[source,xml,subs=attributes+]
+----
+<dependency>
+    <groupId>io.quarkiverse.proxy-wasm</groupId>
+    <artifactId>quarkus-proxy-wasm</artifactId>
+    <version>{project-version}</version>
+</dependency>
+----
+
+[[extension-configuration-reference]]
+== Extension Configuration Reference
+
+TIP: Remove this section if you don't have Quarkus configuration properties in your extension.
+
+include::includes/quarkus-proxy-wasm.adoc[leveloffset=+1, opts=optional]
