+++ quarkus-proxy-wasm/docs/templates/includes/attributes.adoc
@@ -1,0 +1,3 @@
+:project-version: ${release.current-version}
+
+:examples-dir: ./../examples/
