+++ quarkus-proxy-wasm/integration-tests/src/test/java/io/quarkiverse/proxywasm/it/FFIIT.java
@@ -1,0 +1,6 @@
+package io.quarkiverse.proxywasm.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class FFIIT extends FFITest {}
