+++ quarkus-proxy-wasm/integration-tests/src/test/java/io/quarkiverse/proxywasm/it/HeadersIT.java
@@ -1,0 +1,6 @@
+package io.quarkiverse.proxywasm.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class HeadersIT extends HeadersTest {}
